refactor: sorting-factory
