+++ src/sorting-factory.js
